+++ CHANGES.md
@@ -11,6 +11,7 @@ Handy Git hooks to integrate with [pre-commit](http://pre-commit.com/) framework
   * Bumped Checkstyle to 10.23.1
   * Corrected `trailing-whitespaces` hook's syntax
   * Updatd `end-of-file` hook docs
+  * Fixed `checkstyle-jar` cache-dir handling when `--jar` is used
 
 
 * v1.3.1 (2022-11-07)